Commit 18072017

+++ MNTR/src/main/java/com/cbo/mntr/dao/UserDAO.java
@@ -6,7 +6,7 @@ public interface UserDAO {
 
 	public UserInfo getUserByName(String userName);
 
-	public Integer getUserCount();
+	public Long getUserCount();
 
 	public void saveSystemUser(UserInfo user);
